Swapping out Avro and Jsonschema test.
Summary of findings so far:

While all 3 tests start out with the correct KafkaRestConfig (e.g. setting 3 bootstrap.servers), the Avro one uses some other value for the ProducerConfig (e.g. single bootstrap server, none of the ones in KafkaRestConfig), while the other two tests use the correct values.
