+++ kafka-rest/src/test/java/io/confluent/kafkarest/integration/v2/JsonSchemaProduceConsumeTest.java
@@ -1,69 +1,77 @@
 package io.confluent.kafkarest.integration.v2;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.IntNode;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.confluent.kafka.schemaregistry.ParsedSchema;
-import io.confluent.kafka.schemaregistry.json.JsonSchema;
+import io.confluent.kafka.schemaregistry.avro.AvroSchema;
 import io.confluent.kafkarest.Versions;
-import io.confluent.kafkarest.converters.JsonSchemaConverter;
+import io.confluent.kafkarest.converters.AvroConverter;
 import io.confluent.kafkarest.entities.EmbeddedFormat;
 import io.confluent.kafkarest.entities.v2.SchemaTopicProduceRequest.SchemaTopicProduceRecord;
-
 import java.util.Arrays;
 import java.util.List;
+import org.apache.avro.Schema;
+import org.apache.avro.generic.GenericRecordBuilder;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 
 public final class JsonSchemaProduceConsumeTest extends SchemaProduceConsumeTest {
 
-  private static final JsonSchemaConverter JSON_SCHEMA_CONVERTER = new JsonSchemaConverter();
+  private static final AvroConverter AVRO_CONVERTER = new AvroConverter();
 
-  private static final JsonSchema KEY_SCHEMA =
-      new JsonSchema("{\"type\":\"number\"}");
+  private static final Schema KEY_SCHEMA =
+      new Schema.Parser().parse(""
+          + "{"
+          + "  \"type\": \"int\","
+          + "  \"name\": \"key\""
+          + "}");
 
-  private static final JsonSchema VALUE_SCHEMA =
-      new JsonSchema("{\"type\":\"object\",\"properties\":{\"value\":{\"type\":\"number\"}}}");
+  private static final Schema VALUE_SCHEMA =
+      new Schema.Parser().parse(""
+          + "{"
+          + "  \"type\": \"record\", "
+          + "  \"name\": \"ValueRecord\","
+          + "  \"fields\":[{"
+          + "    \"name\": \"value\", "
+          + "    \"type\": \"int\""
+          + "  }]"
+          + "}");
 
   private static final List<SchemaTopicProduceRecord> PRODUCE_RECORDS =
       Arrays.asList(
           new SchemaTopicProduceRecord(
               new IntNode(1),
-              JSON_SCHEMA_CONVERTER.toJson(getMessage(11)).getJson(),
+              AVRO_CONVERTER.toJson(
+                  new GenericRecordBuilder(VALUE_SCHEMA).set("value", 11).build()).getJson(),
               /* partition= */ 0),
           new SchemaTopicProduceRecord(
               new IntNode(2),
-              JSON_SCHEMA_CONVERTER.toJson(getMessage(12)).getJson(),
+              AVRO_CONVERTER.toJson(
+                  new GenericRecordBuilder(VALUE_SCHEMA).set("value", 12).build()).getJson(),
               /* partition= */ 0),
           new SchemaTopicProduceRecord(
               new IntNode(3),
-              JSON_SCHEMA_CONVERTER.toJson(getMessage(13)).getJson(),
+              AVRO_CONVERTER.toJson(
+                  new GenericRecordBuilder(VALUE_SCHEMA).set("value", 13).build()).getJson(),
               /* partition= */ 0));
-
-  private static JsonNode getMessage(int value) {
-    ObjectMapper objectMapper = new ObjectMapper();
-    ObjectNode root = objectMapper.createObjectNode();
-    root.put("value", value);
-    return root;
-  }
 
   @Override
   protected EmbeddedFormat getFormat() {
-    return EmbeddedFormat.JSONSCHEMA;
+    return EmbeddedFormat.AVRO;
   }
 
   @Override
   protected String getContentType() {
-    return Versions.KAFKA_V2_JSON_JSON_SCHEMA;
+    return Versions.KAFKA_V2_JSON_AVRO;
   }
 
   @Override
   protected ParsedSchema getKeySchema() {
-    return KEY_SCHEMA;
+    return new AvroSchema(KEY_SCHEMA);
   }
 
   @Override
   protected ParsedSchema getValueSchema() {
-    return VALUE_SCHEMA;
+    return new AvroSchema(VALUE_SCHEMA);
   }
 
   @Override
